add GraphParse.js, add filterByDate.js, rearrange LineGraphButton
