+++ src/Components/LineGraphButton.js
@@ -1,40 +1,29 @@
 import React, { useState } from "react";
 import LineGraph from "./LineGraph/LineGraph";
-import Graph from "./Graph";
+import GraphParse from "./GraphParse";
 import ChoroplethParent from "../Components/ChoroplethMap/ChoroplethParent";
 
 import "./scss/lineGraphButton.scss";
 
 const LineGraphButton = props => {
-  const { data, chartData, filters, queryType, open, setOpen } = props;
-  const graphItems = filters[1].selectedTableColumnName !== "";
-  console.log(data);
-  const buttonBar = e => {
-    e.preventDefault();
-    setOpen("bar");
-    renderBar();
-  };
+  const {
+    data,
+    filters,
+    open,
+    setOpen,
+    queryType,
+    filterBoxStartDate,
+    filterBoxEndDate
+  } = props;
+  console.log(`data`, data);
 
-  const buttonChoroMap = e => {
-    e.preventDefault();
-    setOpen("choropleth");
-  };
-  const buttonDotMap = e => {
-    e.preventDefault();
-    setOpen("dot");
-  };
+  console.log(data);
 
   const renderLine = () => {
     if (open === "line") {
       return (
         <>
-          <LineGraph
-            filter0={filters[0]}
-            buttonBar={buttonBar}
-            buttonChoroMap={buttonChoroMap}
-            buttonDotMap={buttonDotMap}
-            data={data}
-          />
+          <LineGraph filter0={filters[0]} data={data} />
         </>
       );
     } else {
@@ -78,35 +67,18 @@ const LineGraphButton = props => {
   };
 
   const renderBar = () => {
-    if (graphItems === true && open === "bar") {
+    if (open === "bar") {
       return (
-        <>
-          <Graph
-            data={chartData.percentageData}
-            csvData={chartData.dataStructure}
-            filters={filters}
-            keys={chartData.crossFilterValues}
-            groupMode={"grouped"}
-            sampleSize={chartData.totalSampleSize}
-            tableName={queryType === "sessionsData" ? "Sessions" : "Users"}
-          />
-        </>
+        <GraphParse
+          data={data}
+          filters={filters}
+          open={open}
+          queryType={queryType}
+          filterBoxStartDate={filterBoxStartDate}
+          filterBoxEndDate={filterBoxEndDate}
+        />
       );
-    } else if (graphItems === false && open === "bar") {
-      return (
-        <>
-          <Graph
-            data={chartData.percentageData}
-            csvData={chartData.dataStructure}
-            filters={filters}
-            keys={chartData.keys || chartData.csvKeys}
-            groupMode={"stacked"}
-            sampleSize={chartData.sampleSize}
-            tableName={queryType === "sessionsData" ? "Sessions" : "Users"}
-          />
-        </>
-      );
-    } else if (open !== "bar") {
+    } else {
       return <button onClick={() => setOpen("bar")}>Display Bar Chart</button>;
     }
   };
@@ -114,8 +86,8 @@ const LineGraphButton = props => {
   if (data.sessionsData) {
     return (
       <>
-        {renderLine()}
         {renderBar()}
+        {renderLine()}
         {renderChoroplethMap()}
         {renderDotMap()}
       </>
